+++ src/Console/Commands/QueueFailedCommand.php
@@ -16,12 +16,6 @@ use Vortex\Queue\FailedJobStore;
  */
 final class QueueFailedCommand extends Command
 {
-    public function __construct(
-        private readonly string $basePath,
-    ) {
-        parent::__construct($basePath);
-    }
-
     public function name(): string
     {
         return 'queue:failed';